+++ src/lapidary/runtime/__init__.py
@@ -1,11 +1,13 @@
 __all__ = (
+    'Body',
     'ClientBase',
+    'Cookie',
+    'Header',
     'ModelBase',
     'NamedAuth',
     'ParamStyle',
-    'RequestBody',
-    'ResponseEnvelope',
-    'ResponseBody',
+    'Path',
+    'Query',
     'Responses',
     'SecurityRequirements',
     'delete',
@@ -15,24 +17,12 @@ __all__ = (
     'post',
     'put',
     'trace',
-    'Cookie',
-    'Header',
-    'Path',
-    'Query',
 )
 
+from .annotations import Body, Cookie, Header, Path, Query
 from .client_base import ClientBase
 from .model import ModelBase
-from .model.annotations import (
-    Cookie,
-    Header,
-    Path,
-    Query,
-    RequestBody,
-    ResponseBody,
-    Responses,
-)
 from .model.encode_param import ParamStyle
-from .model.response import ResponseEnvelope
+from .model.response import Responses
 from .operation import delete, get, head, patch, post, put, trace
 from .types_ import NamedAuth, SecurityRequirements