+++ src/lapidary/runtime/response.py
@@ -1,0 +1,8 @@
+import pydantic
+import typing_extensions as typing
+
+from .annotations import StatusCode
+
+
+class ResponseHeaders(pydantic.BaseModel):
+    status_code: typing.Annotated[int, StatusCode]